requirements.txt
folium==0.12.1.post1
geocoder==1.38.1
keras==2.9.0
matplotlib==3.5.2
moviepy==1.0.3
numpy==1.23.0
opencv_python==4.6.0.66
pafy==0.5.5
pandas==1.4.3
Pillow==9.1.1
Pyrebase4==4.5.0
scikit_learn==1.1.1
streamlit==1.10.0
streamlit_option_menu==0.3.2
tensorflow==2.9.1
youtube_dl==2021.12.17
